+++ tests/SEIR.R
@@ -349,9 +349,9 @@ p_expected <- structure(
     .Names = c("time", "prevalence"),
     row.names = c(NA, -10L),
     class = "data.frame")
-p_observed <- prevalence(result, I ~ S + E + I + R, type = "nop")
+p_observed <- prevalence(result, I ~ S + E + I + R, level = 2)
 stopifnot(identical(p_observed, p_expected))
-p_observed <- prevalence(result, I ~ ., type = "nop")
+p_observed <- prevalence(result, I ~ ., level = 2)
 stopifnot(identical(p_observed, p_expected))
 
 p_expected <- data.frame(
@@ -382,11 +382,11 @@ p_expected <- data.frame(
                    0.252886836027714, 0.25234521575985, 0.251974723538705,
                    0.259328358208955, 0.25534188034188, 0.25374251497006,
                    0.252880184331797, 0.252340823970037, 0.251971608832808))
-p_observed <- prevalence(result, I ~ ., type = "wnp")
+p_observed <- prevalence(result, I ~ ., level = 3)
 stopifnot(identical(p_observed$node, p_expected$node))
 stopifnot(identical(p_observed$time, p_expected$time))
 stopifnot(all(abs(p_observed$prevalence - p_expected$prevalence) < tol))
-p_observed <- prevalence(result, I ~ S + E + I + R, type = "wnp")
+p_observed <- prevalence(result, I ~ S + E + I + R, level = 3)
 stopifnot(identical(p_observed$node, p_expected$node))
 stopifnot(identical(p_observed$time, p_expected$time))
 stopifnot(all(abs(p_observed$prevalence - p_expected$prevalence) < tol))
@@ -402,11 +402,11 @@ p_expected <- data.frame(
                    0.255411255411255, 0.253776435045317, 0.255387931034483,
                    0.253765060240964, 0.255364806866953, 0.253753753753754,
                    0.25534188034188, 0.25374251497006))
-p_observed <- prevalence(result, I~., type = "wnp", i = 2:3)
+p_observed <- prevalence(result, I~., level = 3, i = 2:3)
 stopifnot(identical(p_observed$node, p_expected$node))
 stopifnot(identical(p_observed$time, p_expected$time))
 stopifnot(all(abs(p_observed$prevalence - p_expected$prevalence) < tol))
-p_observed <- prevalence(result, I ~ S + E + I + R, type = "wnp", i = 2:3)
+p_observed <- prevalence(result, I ~ S + E + I + R, level = 3, i = 2:3)
 stopifnot(identical(p_observed$node, p_expected$node))
 stopifnot(identical(p_observed$time, p_expected$time))
 stopifnot(all(abs(p_observed$prevalence - p_expected$prevalence) < tol))